Initial Commit
This is the baseline version of my Web Scraper after I split it into multiple modules. First time working with Rust. This version can copy the contents of x amount of posts from x amount of subreddits and uses some multithreading (room to improve for large batches but unnecessary for current needs.